pyproject.toml
[project]
dependencies = [
    "Flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "waitress>=2.1.2",
    "Flask-SQLAlchemy>=3.1.1",
    "psycopg2-binary>=2.9.9",
    "Flask-Migrate>=4.0.0",
    "Flask-Marshmallow>=0.15.0",
    "marshmallow-sqlalchemy>=0.29.0",
]
name = "cursus"
version = "1.0.0"
description = "An open-source API repository for university courses."
license = { file = "LICENSE" }
authors = [{ name = "Rafael Cenzano", email = "richard@richardhnguyen.com" }]

[project.scripts]
cursus = "manage:cli"

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

# For black formatter
[tool.black]
line-length = 79
target-version = ["py311"]
include = '\.pyi?$'
